refactor: remove redundant processed data wrappers


Copilot-Session: e047076f-f61b-47ba-a59b-076859f488e4
