+++ __tests__/helpers/processCSVData.ts
@@ -1,1 +1,2 @@
-export { analyzeData, processCSVData } from '@/utils/analytics/transformations';
+export { analyzeData } from '@/utils/analytics/transformations';
+export { buildProcessedDataFromRawRows as processCSVData } from '@/utils/ingestion';
