+++ src/utils/ingestion/adapter.ts
@@ -1,21 +1,4 @@
-/**
- * Bridge adapter for backward compatibility.
- * Converts new aggregator outputs back to legacy ProcessedData[] format
- * while components are being migrated.
- */
-
-import { CSVData, ProcessedData } from '@/types/csv';
-
-import { buildProcessedDataFromRawRows } from './adapters';
 import { QuotaArtifacts, UsageArtifacts } from './types';
-
-/**
- * Temporary: reconstruct ProcessedData[] from raw CSV for components
- * that haven't been migrated to use aggregator outputs directly.
- */
-export function buildProcessedDataLegacy(rawData: CSVData[]): ProcessedData[] {
-  return buildProcessedDataFromRawRows(rawData);
-}
 
 /**
  * Enriches UserAggregate[] with quota values from QuotaArtifacts.
